Add lower-level bytecode compiler and interpreter
