+++ .travis.yml
@@ -6,7 +6,7 @@ jdk:
   - openjdk8
 
 script:
-  - curl -L -o ~/bin/mill https://github.com/lihaoyi/mill/releases/download/0.2.6/0.2.6 && chmod +x ~/bin/mill
+  - curl -L -o ~/bin/mill https://github.com/lihaoyi/mill/releases/download/0.2.7/0.2.7 && chmod +x ~/bin/mill
   - export PATH=~/bin/mill:$PATH
   - mill all {core,runtime}.test
