+++ runtime/resources/reference.conf
@@ -20,6 +20,11 @@ swam {
   # Configuration that only affects how the runtime part behaves.
   runtime {
 
+    # Whether the engine compiles and run low-level assembly language.
+    # This language aims to be more efficient but is farther from the
+    # bytecode and may be harder to debug.
+    use-low-level-asm = false
+
     stack {
 
       # The height of the local execution stack
