+++ runtime/src/swam/runtime/StackFrame.scala
@@ -16,28 +16,8 @@
 
 package swam
 package runtime
-package config
 
-import com.typesafe.config.Config
-
-class ConfigEngineConfiguration(wrapped: Config) extends EngineConfiguration {
-
-  def useLowLevelAsm = wrapped.getBoolean("swam.runtime.use-low-level-asm")
-
-  object stack extends StackConfiguration {
-
-    def height = wrapped.getInt("swam.runtime.stack.size")
-
-    def callDepth = wrapped.getInt("swam.runtime.stack.call-depth")
-
-  }
-
-  object data extends DataConfiguration {
-
-    def onHeap = wrapped.getBoolean("swam.runtime.data.on-heap")
-
-    def hardMax = wrapped.getInt("swam.runtime.data.hard-max")
-
-  }
-
-}
+/** Interface of a stack frame, making it possible to create
+  * a call stack on trap for reporting purpose.
+  */
+trait StackFrame {}
